# cmd/api/static/openapi.yaml
openapi: 3.0.3
info:
  title: DriveBai API
  description: |
    DriveBai - Car Marketplace and Rental App API

    ## Authentication
    Most endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Error Handling
    All errors return a JSON object with an `error` field containing `code` and `message`.

    Error codes:
    - `EMAIL_TAKEN` - Email already registered
    - `INVALID_CREDENTIALS` - Wrong email or password
    - `RATE_LIMITED` - Too many requests
    - `UNAUTHORIZED` - Authentication required
    - `TOKEN_EXPIRED` - Access token expired
    - `TOKEN_INVALID` - Invalid token
    - `INVALID_ROLE` - Invalid role specified
    - `INCOMPLETE_ONBOARDING` - Onboarding requirements not met
  version: 0.2.0
  contact:
    name: DriveBai Support
    email: support@drivebai.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.drivebai.com/api/v1
    description: Production server
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: |
        Creates a new user account and returns auth tokens immediately.
        User is logged in right after registration.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created and logged in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email (deprecated)
      description: |
        **DEPRECATED** - Email verification is no longer required.
        This endpoint returns success for backwards compatibility.
      deprecated: true
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      description: Authenticates a user and returns access and refresh tokens.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/token/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: |
        Exchanges a refresh token for a new access token and refresh token.
        The old refresh token is invalidated (token rotation).
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/password/forgot:
    post:
      tags: [Auth]
      summary: Request password reset
      description: |
        Sends a password reset link to the user's email.
        Always returns 202 to prevent email enumeration.
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '202':
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/password/reset:
    post:
      tags: [Auth]
      summary: Reset password with token
      description: |
        Resets the user's password using the token from the reset link.
        All existing sessions are invalidated.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout and revoke refresh token
      description: Revokes the provided refresh token.
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/resend-otp:
    post:
      tags: [Auth]
      summary: Resend OTP code (deprecated)
      description: |
        **DEPRECATED** - OTP is no longer used.
        This endpoint returns success for backwards compatibility.
      deprecated: true
      operationId: resendOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendOTPRequest'
      responses:
        '202':
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /me:
    get:
      tags: [User]
      summary: Get current user profile
      description: Returns the authenticated user's profile information.
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profile:
    patch:
      tags: [User]
      summary: Update user profile
      description: |
        Updates the authenticated user's profile information.
        All fields are optional - only provided fields will be updated.
        Role can only be set to 'driver' or 'car_owner'.
      operationId: updateProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateProfileResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profile/photo:
    post:
      tags: [User]
      summary: Upload profile photo
      description: Uploads a profile photo for the authenticated user.
      operationId: uploadProfilePhoto
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (JPEG, PNG)
      responses:
        '200':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  profile_photo_url:
                    type: string
        '400':
          description: Invalid file type or size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents:
    get:
      tags: [Documents]
      summary: Get user's documents
      description: Returns all documents uploaded by the authenticated user.
      operationId: getDocuments
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/{type}:
    post:
      tags: [Documents]
      summary: Upload a document
      description: |
        Uploads a document of the specified type.
        If a document of the same type already exists, it will be replaced.
      operationId: uploadDocument
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [drivers_license, registration]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Document file (JPEG, PNG, PDF)
      responses:
        '201':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Invalid file type or document type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/{id}:
    delete:
      tags: [Documents]
      summary: Delete a document
      description: Deletes the specified document.
      operationId: deleteDocument
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /onboarding/complete:
    post:
      tags: [Onboarding]
      summary: Complete onboarding
      description: |
        Marks the user's onboarding as complete.
        For drivers, required documents must be uploaded first.
      operationId: completeOnboarding
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Onboarding completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Onboarding requirements not met
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - first_name
        - last_name
        - role
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: SecurePass123
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        phone:
          type: string
          example: "+1234567890"
        role:
          type: string
          enum: [driver, car_owner]
          example: driver
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/UserProfile'
    VerifyEmailRequest:
      type: object
      required:
        - email
        - code
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          pattern: ^\d{6}$
          example: "123456"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required:
        - token
        - new_password
      properties:
        token:
          type: string
          description: Token from password reset email
        new_password:
          type: string
          format: password
          minLength: 8
    ResendOTPRequest:
      type: object
      required:
        - email
        - purpose
      properties:
        email:
          type: string
          format: email
        purpose:
          type: string
          enum: [verify_email, reset_password]
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [driver, car_owner, admin]
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
          nullable: true
        is_email_verified:
          type: boolean
        onboarding_status:
          type: string
          enum: [created, role_selected, photo_uploaded, documents_uploaded, complete]
        profile_photo_url:
          type: string
          nullable: true
    UpdateProfileRequest:
      type: object
      properties:
        role:
          type: string
          enum: [driver, car_owner]
          description: User role (cannot be set to admin via API)
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
          nullable: true
    UpdateProfileResponse:
      type: object
      properties:
        message:
          type: string
        data:
          $ref: '#/components/schemas/UserProfile'
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [drivers_license, registration]
        file_name:
          type: string
        file_size:
          type: integer
        status:
          type: string
          enum: [uploaded, verified, rejected]
        created_at:
          type: string
          format: date-time
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_INPUT
            message:
              type: string
              example: Email is required